+++ util/array/flatten.ts
@@ -9,34 +9,36 @@
  * <BR>
  *
  * ```javascript
- * import { arrayFlatten } from 'https://deno.land/x/polkadot@0.2.40/util/mod.ts';
+ * import { arrayFlatten } from 'https://deno.land/x/polkadot/util/mod.ts';
  *
  * arrayFlatten([[1, 2], [3, 4], [5]]); // [1, 2, 3, 4, 5]
  * ```
  */
 export function arrayFlatten <T> (arrays: readonly T[][]): T[] {
+  const num = arrays.length;
+
   // shortcuts for the empty & single-entry case
-  if (arrays.length === 0) {
+  if (num === 0) {
     return [];
-  } else if (arrays.length === 1) {
+  } else if (num === 1) {
     return arrays[0];
   }
 
   // pre-allocate based on the combined size
   let size = 0;
 
-  for (let i = 0; i < arrays.length; i++) {
+  for (let i = 0; i < num; i++) {
     size += arrays[i].length;
   }
 
   const output = new Array<T>(size);
   let i = -1;
 
-  for (let j = 0; j < arrays.length; j++) {
+  for (let j = 0; j < num; j++) {
     const a = arrays[j];
 
     // instead of pushing, we just set the entries
-    for (let e = 0; e < a.length; e++) {
+    for (let e = 0, count = a.length; e < count; e++) {
       output[++i] = a[e];
     }
   }