+++ util/float/toU8a.ts
@@ -1,6 +1,4 @@
 
-type AnyFloat = number | Number;
-
 interface Options {
   bitLength?: 32 | 64;
   isLe?: boolean;
@@ -11,7 +9,7 @@ interface Options {
  * @description Converts a float into a U8a representation (While we don't use BE in SCALE
  * we still allow for either representation, although, as elsewhere, isLe is default)
  */
-export function floatToU8a (value: string | AnyFloat = 0.0, { bitLength = 32, isLe = true }: Options = {}): Uint8Array {
+export function floatToU8a (value: String | string | number | Number = 0.0, { bitLength = 32, isLe = true }: Options = {}): Uint8Array {
   if (bitLength !== 32 && bitLength !== 64) {
     throw new Error('Invalid bitLength provided, expected 32 or 64');
   }