+++ util/hex/toU8a.ts
@@ -5,7 +5,7 @@ const CHR = '0123456789abcdef';
 const U8 = new Uint8Array(256);
 const U16 = new Uint8Array(256 * 256);
 
-for (let i = 0; i < CHR.length; i++) {
+for (let i = 0, count = CHR.length; i < count; i++) {
   U8[CHR[i].charCodeAt(0) | 0] = i | 0;
 
   if (i > 9) {
@@ -30,7 +30,7 @@ for (let i = 0; i < 256; i++) {
  * <BR>
  *
  * ```javascript
- * import { hexToU8a } from 'https://deno.land/x/polkadot@0.2.40/util/mod.ts';
+ * import { hexToU8a } from 'https://deno.land/x/polkadot/util/mod.ts';
  *
  * hexToU8a('0x80001f'); // Uint8Array([0x80, 0x00, 0x1f])
  * hexToU8a('0x80001f', 32); // Uint8Array([0x00, 0x80, 0x00, 0x1f])