+++ util/is/function.ts
@@ -10,7 +10,7 @@ type FnType = Function;
  * <BR>
  *
  * ```javascript
- * import { isFunction } from 'https://deno.land/x/polkadot@0.2.40/util/mod.ts';
+ * import { isFunction } from 'https://deno.land/x/polkadot/util/mod.ts';
  *
  * isFunction(() => false); // => true
  * ```