+++ util/promisify.ts
@@ -14,7 +14,7 @@
  * await promisify(null, (cb) => cb(new Error('error!'))); // rejects with `error!`
  * ```
  */
-export function promisify <R = any> (self: unknown, fn: (...params: any) => any, ...params: any[]): Promise<R> {
+export function promisify <R = any> (self: unknown, fn: (...params: any[]) => any, ...params: any[]): Promise<R> {
   return new Promise((resolve, reject): void => {
     fn.apply(self, params.concat((error: Error | null, result?: R): void => {
       if (error) {