+++ util/types.ts
@@ -69,8 +69,8 @@ export type HexDigit = '0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9'
 export type HexString = `0x${string}`;
 
 export interface BufferObj extends Uint8Array {
-  equals (otherBuffer: Uint8Array): boolean;
-  readDoubleLE (offset?: number): number;
+  equals: (otherBuffer: Uint8Array) => boolean;
+  readDoubleLE: (offset?: number) => number;
 }
 
 export interface BufferObjConstructor extends Constructor<BufferObj> {